+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased dev version
+
+### Added
+* API support to surrender individual tasks.
+
+### Removed
+* Support for Python 3.10.
+
+### Changed
+* reorganisation of some Marking and Identifying tools in the UI.
+
+### Fixed
+
+
+
+
 ## [Unreleased]
 
 ### Added